backend/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

psycopg2-binary==2.9.9
opensearch-py==2.6.0
redis==5.0.4

# Pydantic v2 + email validator
pydantic[email]==2.7.4
email-validator==2.1.0.post1

# Auth & upload
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
python-jose==3.3.0
python-multipart==0.0.9
PyJWT==2.8.0
